+++ app/routes/components/Select/changelog.codegen.ts
@@ -1,3 +1,4 @@
 export const changelog = ''
   + '- 2025-09-23: Initial version.\n'
+  + '- 2025-09-25: Allow props to accept generics.\n'
   ;